Remove using the Loader for logger
